Compile with 64bit integer interface
  Add option to complie with both MKL ilp64 and lp64 interfaces as
  well as BLA_SIZEOF_INTEGER set to 8 or 4. The compiler flag to
  set default integers to 8/4 shuld not cause issues regardless of
  the chosen BLAS interface since the blas_int kind is set only
  in linalg_wrapper.f90. Compilers will throw conversion warnings
  in that module when the BLAS interface is different than the
  default integer, but the integers should never be large enough to
  cause an issue.

  Fixed error thrown by random.f90 with -i8 or -fdefault-integer-8
  flags.

+++ src/util/random.f90
@@ -189,7 +189,7 @@ contains
     !----------------------------------------------------------------------------------------------
     subroutine rng_integer_1d(self, rnum)
         class(rng_type) :: self
-        integer, intent(out) :: rnum(:)
+        integer(int32), intent(out) :: rnum(:)
         integer :: i
 
         do i = 1, size(rnum)
@@ -204,7 +204,7 @@ contains
     !----------------------------------------------------------------------------------------------
     subroutine rng_integer_2d(self, rnum)
         class(rng_type) :: self
-        integer, intent(out) :: rnum(:, :)
+        integer(int32), intent(out) :: rnum(:, :)
         integer :: i
 
         do i = 1, size(rnum, 2)
@@ -219,7 +219,7 @@ contains
     !----------------------------------------------------------------------------------------------
     subroutine rng_integer_3d(self, rnum)
         class(rng_type) :: self
-        integer, intent(out) :: rnum(:, :, :)
+        integer(int32), intent(out) :: rnum(:, :, :)
         integer :: i
 
         do i = 1, size(rnum, 3)
